code_tr/requirements.txt
diffusers==0.11.1
transformers==4.25.1
einops==0.8.1
opencv-python==4.10.0.84
Pillow==10.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
huggingface-hub==0.17.3
xformers==0.0.16
pyre-extensions==0.0.23
